requirements.txt
ftfy
regex
tqdm
gdown
scipy
transformers==4.35.2
matplotlib
open_clip_torch==2.23.0
numpy
pandas
einops
flake8==3.7.9
yapf==0.29.0
isort==4.3.21
yacs
tb-nightly
future
scikit-learn
wilds==1.2.2
tabulate